JavadocAppenderLogicおよびJavadocAppenderLogicImplクラスを削除し、JdtsReplLogicおよびJdtsReplLogicImplクラスに関連するロジックを追加。Javadocタグ設定の処理を統一し、コードの可読性とメンテナンス性を向上。
